+++ Marchgame/org.opengl.setupclasses/texture/TextureHolder.java
@@ -8,28 +8,28 @@ import org.lwjgl.util.vector.Vector2f;
 import utils.TextureUtils;
 
 public class TextureHolder {
-	private int parentwidth;
-	private int parentheight;
+	private float parentwidth;
+	private float parentheight;
 	private ArrayList<Texture> tex = new ArrayList<Texture>();
 	private int currenttexid = 0;
 	private BufferedImage img;
 	
-	public TextureHolder(int parentsize, BufferedImage img) {
+	public TextureHolder(float parentsize, BufferedImage img) {
 		this.parentwidth = parentsize;
 		this.parentheight = parentsize;
 		this.img = img;
 	}
-	public TextureHolder(int parentwidth, int parentheight, BufferedImage img) {
+	public TextureHolder(float parentwidth, float parentheight, BufferedImage img) {
 		this.parentwidth = parentwidth;
 		this.parentheight = parentheight;
 		
 		TextureUtils util = new TextureUtils();
 	}
-	public void addTexture(Vector2f topleft, int size) {
+	public void addTexture(Vector2f topleft, float size) {
 		tex.add(new Texture(currenttexid, topleft, size, parentwidth));
 		currenttexid += 1;
 	}
-	public void addTexture(Vector2f topleft, int width, int height) {
+	public void addTexture(Vector2f topleft, float width, float height) {
 		tex.add(new Texture(currenttexid, topleft, width, height, parentwidth, parentheight));
 		currenttexid += 1;
 	}
